Refactor, add tests
