+++ package.json
@@ -1,14 +1,20 @@
 {
-    "name": "pdf-thumbnails",
-    "version": "1.0.0",
-    "main": "src/app.js",
-    "repository": "https://www.github.com/tind/pdf-thumbnails",
-    "license": "MIT",
-    "dependencies": {
-	"canvas": "^2.6.1",
-	"pdfjs-dist": "^2.5.207"
-    },
-    "bin": {
-	"pdfthumbnail": "./src/app.js"
-    }
+   "name": "pdf-thumbnails",
+   "version": "1.0.0",
+   "main": "src/app.js",
+   "repository": "https://www.github.com/tind/pdf-thumbnails",
+   "license": "MIT",
+   "dependencies": {
+      "canvas": "^2.6.1",
+      "pdfjs-dist": "^2.5.207"
+   },
+   "devDependencies": {
+      "mocha": "^8.2.1"
+   },
+   "scripts": {
+      "test": "mocha"
+   },
+   "bin": {
+      "pdfthumbnail": "./src/app.js"
+   }
 }
